Changed way to mute the video
I changed the way the video/ad is muted, Before it clicked on the volume button to mute it. And it showed the animation of muting. The problem is that if user opened multiple pages and one of them is showing an ad, and it muted, It would mute all the other pages as well. To avoid it, I changed the way the mute works, instead of clicking the the volume button, it just mutes by JavaScript. With that we avoid muting multiple videos, But the bad news is that there is no longer a mute indicator.